+++ src/pretalx/schedule/models/room.py
@@ -8,43 +8,48 @@ from pretalx.common.urls import EventUrls
 
 class Room(LogMixin, models.Model):
     event = models.ForeignKey(
-        to='event.Event',
-        on_delete=models.PROTECT,
-        related_name='rooms',
+        to='event.Event', on_delete=models.PROTECT, related_name='rooms'
     )
-    name = I18nCharField(
-        max_length=100,
-        verbose_name=_('Name'),
-    )
+    name = I18nCharField(max_length=100, verbose_name=_('Name'))
     description = I18nCharField(
         max_length=1000,
-        null=True, blank=True,
+        null=True,
+        blank=True,
         verbose_name=_('Description'),
         help_text=_('A description for attendees, for example directions.'),
     )
     speaker_info = I18nCharField(
         max_length=1000,
-        null=True, blank=True,
+        null=True,
+        blank=True,
         verbose_name=_('Speaker Information'),
-        help_text=_('Information relevant for speakers scheduled in this room, for example room size, special directions, available adapters for video input …'),
+        help_text=_(
+            'Information relevant for speakers scheduled in this room, for example room size, special directions, available adapters for video input …'
+        ),
     )
     capacity = models.PositiveIntegerField(
-        null=True, blank=True,
+        null=True,
+        blank=True,
         verbose_name=_('Capacity'),
-        help_text=_('How many people can fit in the room?')
+        help_text=_('How many people can fit in the room?'),
     )
     position = models.PositiveIntegerField(
-        null=True, blank=True,
+        null=True,
+        blank=True,
         verbose_name=_('Position'),
-        help_text=_('This is the order that rooms are displayed in in the schedule (lower = left).'),
+        help_text=_(
+            'This is the order that rooms are displayed in in the schedule (lower = left).'
+        ),
     )
 
     class Meta:
-        ordering = ('position', )
+        ordering = ('position',)
 
     class urls(EventUrls):
         settings_base = edit = '{self.event.orga_urls.room_settings}/{self.pk}'
         delete = '{settings_base}/delete'
+        up = '{settings_base}/up'
+        down = '{settings_base}/down'
 
     def __str__(self) -> str:
         return f'Room(event={self.event.slug}, name={self.name})'